futwiz-project/pyproject.toml
[tool.poetry]
name = "fut-wiz"
version = "0.1.0"
description = "Project will be to create an API for FUT Wiz."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fut_wiz"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
selenium = "^4.8.0"
python-dotenv = "^0.21.1"
webdriver-manager = "^3.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
